+++ src/main/java/com/cedarsoftware/util/CaseInsensitiveMap.java
@@ -1,6 +1,16 @@
 package com.cedarsoftware.util;
 
-import java.util.*;
+import java.util.AbstractMap;
+import java.util.AbstractSet;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.Iterator;
+import java.util.LinkedHashMap;
+import java.util.LinkedHashSet;
+import java.util.Map;
+import java.util.Set;
+import java.util.TreeMap;
+import java.util.WeakHashMap;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.ConcurrentSkipListMap;
@@ -270,16 +280,16 @@ public class CaseInsensitiveMap<K, V> implements Map<K, V>
     }
 
     /**
-     * Returns a {@link Set} view of the keys contained in this map.
+     * Returns a Set view of the keys contained in this map.
      * The set is backed by the map, so changes to the map are
      * reflected in the set, and vice-versa.  If the map is modified
      * while an iteration over the set is in progress (except through
-     * the iterator's own <tt>remove</tt> operation), the results of
+     * the iterator's own <b>remove</b> operation), the results of
      * the iteration are undefined.  The set supports element removal,
      * which removes the corresponding mapping from the map, via the
-     * <tt>Iterator.remove</tt>, <tt>Set.remove</tt>,
-     * <tt>removeAll</tt>, <tt>retainAll</tt>, and <tt>clear</tt>
-     * operations.  It does not support the <tt>add</tt> or <tt>addAll</tt>
+     * <b>Iterator.remove</b>, <b>Set.remove</b>,
+     * <b>removeAll</b>, <b>retainAll</b>, and <b>clear</b>
+     * operations.  It does not support the <b>add</b> or <b>addAll</b>
      * operations.
      */
     public Set<K> keySet()
@@ -478,7 +488,7 @@ public class CaseInsensitiveMap<K, V> implements Map<K, V>
                 return false;
             }
 
-            Map.Entry that = (Map.Entry) o;
+            Map.Entry<K, V> that = (Map.Entry) o;
             if (localMap.containsKey(that.getKey()))
             {
                 Object value = localMap.get(that.getKey());