+++ src/main/java/com/cedarsoftware/util/ExceptionUtilities.java
@@ -28,7 +28,7 @@ public final class ExceptionUtilities
     /**
      * Safely Ignore a Throwable or rethrow if it is a Throwable that should
      * not be ignored.
-     * @param t
+     * @param t Throwable to possibly ignore (ThreadDeath and OutOfMemory are not ignored).
      */
     public static void safelyIgnoreException(Throwable t)
     {