+++ src/test/java/com/cedarsoftware/util/TestCaseInsensitiveMap.java
@@ -1271,7 +1271,7 @@ public class TestCaseInsensitiveMap
 //
 //        start = System.nanoTime();
 //
-//        for (int i=0; i < 10000; i++)
+//        for (int i=0; i < 100000; i++)
 //        {
 //            Map copy = new CaseInsensitiveMap<>(map);
 //        }